+++ example/lib/example.dart
@@ -2,7 +2,7 @@ import 'package:flexi_formatter/flexi_formatter.dart';
 
 void main() {
   /// 987654321.123456789 => '987654321.12345'
-  print(formatNumber(987654321.123456789.d, precision: 5, mode: RoundMode.floor));
+  print(formatNumber(987654321.123456789.d, precision: 5, roundMode: RoundMode.floor));
 
   /// 1.00000123000 => '1.0{5}123'
   print(formatNumber(
@@ -21,22 +21,48 @@ void main() {
   /// 9876543210.1 => '9.88B'
   print(formatAmount(9876543210.1.d));
 
+  /// 9876543210.1 => '9.87B'
+  print(formatAmount(9876543210.1.d, roundMode: RoundMode.floor));
+
   /// 9876543210.1 => '98.77亿'
   print(formatAmount(
     9876543210.1.d,
-    compactConverter: FormatDecimal.simplifiedChineseCompactConverter,
+    compactConverter: FlexiFormatter.simplifiedChineseCompactConverter,
   ));
 
-  /// 123456.000000789 => '$+123 456.0₆789USDT'
-  FormatDecimal.thousandSeparator = ' ';
+  /// 1234567890.000000789 => '￥+1_2345_6789.0₆78元'
   print(formatNumber(
-    123456.000000789.d,
-    precision: 100,
+    '123456789.000000789'.d,
+    precision: 8,
+    roundMode: RoundMode.floor,
     cutInvalidZero: true,
-    showThousands: true,
+    enableGrouping: true,
+    groupSepartor: '_',
+    groupCounts: 4,
     shrinkZeroMode: ShrinkZeroMode.subscript,
-    prefix: '\$',
-    suffix: 'USDT',
     showSign: true,
+    prefix: '￥',
+    suffix: '元',
   ));
+
+  /// 123456789.000000789 => '￥+1.2345.6789,0₆78元'
+  try {
+    FlexiFormatter.setGlobalConfig(decimalSeparator: ',', groupSeparator: '.', groupCounts: 4);
+    print(formatNumber(
+      '123456789.000000789'.d,
+      precision: 8,
+      roundMode: RoundMode.floor,
+      cutInvalidZero: true,
+      enableGrouping: true,
+      shrinkZeroMode: ShrinkZeroMode.subscript,
+      showSign: true,
+      prefix: '￥',
+      suffix: '元',
+    ));
+  } finally {
+    FlexiFormatter.restoreGlobalDefaultConfig();
+  }
+
+  print('￥+123456.0{6}789元'.ltr);
+  print('￥${'+123456.0{6}789'.ltr}元'.rtl);
 }