+++ lib/src/decimal_ext.dart
@@ -14,48 +14,9 @@
 
 part of 'decimal_format_util.dart';
 
-// 上角标的数字符号12345⁰¹²³⁴⁵⁶⁷⁸⁹⁺⁻
-const superscriptNumerals = ['⁰', '¹', '²', '³', '⁴', '⁵', '⁶', '⁷', '⁸', '⁹'];
-
-// 上角标的正负号
-const superscriptPositive = '⁺';
-const superscriptNegative = '⁻';
-
-// 下角标的数字符号12345₀₁₂₃₄₅₆₇₈₉₊₋
-const subscriptNumerals = ['₀', '₁', '₂', '₃', '₄', '₅', '₆', '₇', '₈', '₉'];
-
-// 下角标的正负号
-const subscriptPositive = '₊';
-const subscriptNegative = '₋';
-
-// 默认Decimal除法精度
-const int defaultScaleOnInfinitePrecision = 17;
-
-typedef CompactConverter = (Decimal, String) Function(Decimal);
-
-enum RoundMode {
-  round,
-  floor,
-  ceil,
-  truncate,
-}
-
-enum UnicodeDirection {
-  ltr,
-  rtl;
-}
-
-enum ShrinkZeroMode {
-  compact, // 0.0000123=>0.0{4}123
-  subscript, // 0.0000123=>0.0₄123
-  superscript; // 0.0000123=>0.0⁴123
-}
-
 extension StringExt on String {
   Decimal? get decimal => Decimal.tryParse(this);
   Decimal get d => Decimal.tryParse(this) ?? Decimal.zero;
-
-  String get cleanedString => cleaned;
 
   /// https://unicode.org/reports/tr9/
   /// 从左到右嵌入
@@ -102,13 +63,7 @@ extension IntExt on int {
   }
 }
 
-final Decimal two = Decimal.fromInt(2);
-final Decimal three = Decimal.fromInt(3);
-final Decimal twentieth = (Decimal.one / Decimal.fromInt(20)).toDecimal();
-final Decimal fifty = Decimal.fromInt(50);
-final Decimal hundred = Decimal.fromInt(100);
-
-extension FormatDecimal on Decimal {
+extension DecimalExt on Decimal {
   Decimal get half => (this / two).toDecimal(
         scaleOnInfinitePrecision: defaultScaleOnInfinitePrecision,
       );
@@ -127,43 +82,16 @@ extension FormatDecimal on Decimal {
     );
   }
 
-  /// Global configuration for decimal point default [decimalSeparator].
-  static String decimalSeparator = '.';
-
-  /// Global configuration for xThousand formatting default [thousandSeparator].
-  static String thousandSeparator = ',';
-
-  /// Global configuration for compact formatting. if null, The [defaultCompactConverter] will be used
-  static CompactConverter? globalCompactConverter;
-
-  /// Default compact conversion
-  static (Decimal, String) defaultCompactConverter(Decimal value) => value.toThousandCompact;
-
-  /// Simplified Chinese compact conversion
-  static (Decimal, String) simplifiedChineseCompactConverter(Decimal value) =>
-      value.toSimplifiedChineseCompact;
-
-  /// Traditional Chinese compact conversion
-  static (Decimal, String) traditionalChineseCompactConverter(Decimal value) =>
-      value.toTraditionalChineseCompact;
-
-  /// Global configuration for displayed as the minimum boundary value of the exponent.
-  static Decimal exponentMinDecimal = Decimal.ten.pow(-15).toDecimal();
-
-  /// Global configuration for displayed as the maximum boundary value of the exponent
-  static Decimal exponentMaxDecimal = Decimal.ten.pow(21).toDecimal();
-
-  /// Format this number with thousands separators
-  /// If [precision] is not specified, it defaults to 3.
-  (String, String) compact({
+  /// Format this number with compact converter.
+  (String, String) _compact({
     int precision = 2,
     bool isClean = true,
-    RoundMode? mode,
+    RoundMode? roundMode,
     CompactConverter? converter,
   }) {
-    converter ??= globalCompactConverter ?? defaultCompactConverter;
+    converter ??= FlexiFormatter.globalCompactConverter;
     var (value, unit) = converter(this);
-    if (mode != null) value = value.toRoundMode(mode, scale: precision);
+    if (roundMode != null) value = value.toRoundMode(roundMode, scale: precision);
     String result = value.toStringAsFixed(precision);
     if (isClean) result = result.cleaned;
     return (result, unit);
@@ -171,23 +99,13 @@ extension FormatDecimal on Decimal {
 
   /// use [RoundMode] to handling Decimal
   Decimal toRoundMode(RoundMode mode, {int? scale}) {
-    Decimal val = this;
     scale ??= 0;
-    switch (mode) {
-      case RoundMode.round:
-        val = val.round(scale: scale);
-        break;
-      case RoundMode.floor:
-        val = val.floor(scale: scale);
-        break;
-      case RoundMode.ceil:
-        val = val.ceil(scale: scale);
-        break;
-      case RoundMode.truncate:
-        val = val.truncate(scale: scale);
-        break;
-    }
-    return val;
+    return switch (mode) {
+      RoundMode.round => round(scale: scale),
+      RoundMode.floor => floor(scale: scale),
+      RoundMode.ceil => ceil(scale: scale),
+      RoundMode.truncate => truncate(scale: scale),
+    };
   }
 
   /// The rational string that correctly represents this number.
@@ -200,13 +118,14 @@ extension FormatDecimal on Decimal {
   /// [precision] stands for fractionDigits
   String formatAsString(
     int precision, {
-    RoundMode? mode,
+    RoundMode? roundMode,
     bool isClean = true,
   }) {
     Decimal val = this;
-    if (mode != null) val = toRoundMode(mode, scale: precision);
+    if (roundMode != null) val = toRoundMode(roundMode, scale: precision);
     String result;
-    if (val.abs() <= exponentMinDecimal || val.abs() > exponentMaxDecimal) {
+    if (val.abs() <= FlexiFormatter.exponentMinDecimal ||
+        val.abs() > FlexiFormatter.exponentMaxDecimal) {
       result = val.toStringAsExponential(precision);
     } else {
       result = val.toStringAsFixed(precision);
@@ -218,18 +137,19 @@ extension FormatDecimal on Decimal {
   /// Parsing [Decimal] to percentage [String].
   String get percentage => '${(this * hundred).toStringAsFixed(2).cleaned}%';
 
-  /// Parsing Decimal to thousands [String].
-  String thousands(
+  /// Format the integer part of Decimal with grouping configuration
+  (String, String) _group(
     int precision, {
-    RoundMode? mode,
+    RoundMode? roundMode,
     bool isClean = true,
-    String? separator,
+    int? groupCounts,
+    String? groupSeparator,
   }) {
     return formatAsString(
       precision,
-      mode: mode,
+      roundMode: roundMode,
       isClean: isClean,
-    ).thousands(separator ?? thousandSeparator);
+    ).grouping(groupSeparator, groupCounts);
   }
 }
 
@@ -338,17 +258,21 @@ extension on Decimal {
 extension on String {
   String get cleaned {
     return switch (this) {
-      String value when value.endsWith(FormatDecimal.decimalSeparator) =>
+      String value when value.endsWith(defaultDecimalSeparator) =>
         value.substring(0, value.length - 1),
-      String value when value.endsWith('0') && contains(FormatDecimal.decimalSeparator) =>
+      String value when value.endsWith('0') && contains(defaultDecimalSeparator) =>
         value.substring(0, value.length - 1).cleaned,
       String value when value.contains('e') => value.replaceAll(RegExp(r'(?<=\.\d*?)0+(?!\d)'), ''),
       _ => this,
     };
   }
 
-  String thousands(String separator) {
-    final dotIndex = indexOf(FormatDecimal.decimalSeparator);
+  (String, String) grouping(String? separator, int? groupCounts) {
+    separator ??= FlexiFormatter.globalGroupIntegerSeparator;
+    groupCounts ??= FlexiFormatter.globalGroupIntegerCounts;
+    groupCounts = groupCounts.clamp(1, 10);
+
+    final dotIndex = indexOf(defaultDecimalSeparator);
     String integerPart, decimalPart;
     if (dotIndex == -1) {
       integerPart = this;
@@ -357,33 +281,51 @@ extension on String {
       integerPart = substring(0, dotIndex);
       decimalPart = substring(dotIndex, length);
     }
-    final regex = RegExp(r'(\d)(?=(\d{3})+(?!\d))');
     final formattedInteger = integerPart.replaceAllMapped(
-      regex,
+      RegExp('(\\d)(?=(\\d{$groupCounts})+(?!\\d))'),
       (Match match) => '${match[1]}$separator',
     );
 
-    return formattedInteger + decimalPart;
+    return (formattedInteger, decimalPart);
   }
 
-  String shrinkZero(ShrinkZeroMode mode) {
-    final dotIndex = lastIndexOf(FormatDecimal.decimalSeparator);
+  String shrinkZero({
+    ShrinkZeroMode? shrinkMode,
+    ShrinkZeroConverter? shrinkConverter,
+  }) {
+    shrinkMode ??= FlexiFormatter.globalShrinkZeroMode;
+    shrinkConverter ??= FlexiFormatter.globalShrinkZeroConverter;
+    // if (shrinkZeroMode == null) return this;
+
+    final dotIndex = lastIndexOf(defaultDecimalSeparator);
     if (dotIndex == -1) return this;
 
     final decimalPart = substring(dotIndex + 1);
-    final regex = RegExp(r'(0{4,})(?=[1-9]|$)');
     final formattedDecimal = decimalPart.replaceAllMapped(
-      regex,
+      RegExp(r'(0{4,})(?=[1-9]|$)'),
       (Match match) {
-        final zeroLen = match[1]!.length;
-        return switch (mode) {
-          ShrinkZeroMode.compact => '0{$zeroLen}',
-          ShrinkZeroMode.subscript => '0${zeroLen.subscriptNumeral}',
-          ShrinkZeroMode.superscript => '0${zeroLen.superscriptNumeral}'
+        final zeroPart = match[1] ?? '';
+        final zeroCounts = zeroPart.length;
+        if (zeroCounts < 1) return zeroPart;
+        return switch (shrinkMode) {
+          ShrinkZeroMode.compact => '0{$zeroCounts}',
+          ShrinkZeroMode.subscript => '0${zeroCounts.subscriptNumeral}',
+          ShrinkZeroMode.superscript => '0${zeroCounts.superscriptNumeral}',
+          _ => shrinkConverter?.call(zeroCounts) ?? zeroPart,
         };
       },
     );
 
-    return '${substring(0, dotIndex)}${FormatDecimal.decimalSeparator}$formattedDecimal';
+    return '${substring(0, dotIndex)}${FlexiFormatter.globalDecimalSeparator}$formattedDecimal';
+  }
+
+  String unicodeDirection(UnicodeDirection? direction) {
+    direction ??= FlexiFormatter.globalUnicodeDirection;
+    if (direction == null) return this;
+
+    return switch (direction) {
+      UnicodeDirection.ltr => ltr,
+      UnicodeDirection.rtl => rtl,
+    };
   }
 }